requirements.txt
Django==4.2.5
PyMySQL==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
boto3==1.28.50
openai==0.28.0
twilio==8.9.0
